requirements.txt
psycopg2-binary
sqlalchemy
ollama
langchain-community
python-dotenv
pyTelegramBotAPI
atlassian-python-api
strip-tags
beautifulsoup4